Notifications

+++ 14-RESTAPI/server/controllers/postController.ts
@@ -1,53 +1,56 @@
 import Post from '../models/Post';
 import { RequestHandler } from 'express';
 import { io } from '../app';
+import AppError from '../models/Error';
 import { getErrors, hasErrors } from '../validation/validators';
 import { deleteFile } from '../util/deleteFile';
-import captainsLog from '../util/captainsLog';
+
+const devErr = 'Do not use without AuthJWT';
 
 const newPost: RequestHandler = async (req, res, next) => {
+  const user = req.user;
+  if (!user) return next(new AppError(403, 'Something went wrong', devErr));
+
   try {
     const { title, content } = req.body;
-    const image = req.file;
+    const        image       = req.file;
 
     const errors = getErrors(req);
     if (hasErrors(errors)) {
-      res.status(422).json(errors);
       if (image) deleteFile(image.path);
-      return;
+      return next(new AppError(422, errors));
     }
 
-    const post = new Post({ title, content, creator: req.user });
+    const post = new Post({ title, content, creator: user._id });
     if (image) post.imgURL = image.path;
     await post.save();
+
+    user.set({ email: 'hidden', friends: [] });
+    post.creator = user;
+
     io.emit('post:update', post); // pushes socket to client
     res.status(201).json(post);
   } catch (error) {
-    captainsLog(5, 'newPost Catch', error);
-    res.status(500).json({ message: 'Unable to submit your post.' });
+    next(new AppError(500, 'Unable to submit your post', error));
   }
 };
 
 const editPost: RequestHandler = async (req, res, next) => {
   try {
-    const { postId } = req.params;
+    const {     postId     } = req.params;
     const { title, content } = req.body;
-    const image = req.file;
+    const       image        = req.file;
 
     const errors = getErrors(req);
     if (hasErrors(errors)) {
-      res.status(422).json(errors);
       if (image) deleteFile(image.path);
-      return;
+      return next(new AppError(422, errors));
     }
 
     const post = await Post.findOne({ _id: postId, creator: req.user });
-    if (!post) {
-      res.status(404).json({ message: 'Post not found.' });
-      return;
-    }
+    if (!post) return next(new AppError(404, 'Post not found'));
 
-    post.title = title;
+    post.title   = title;
     post.content = content;
     if (image) {
       if (post.imgURL) deleteFile(post.imgURL);
@@ -60,8 +63,7 @@ const editPost: RequestHandler = async (req, res, next) => {
     io.emit(`post:${postId}:update`, post); // emits to specfic path only
     res.status(200).json(post);
   } catch (error) {
-    captainsLog(5, 'editPost Catch', error);
-    res.status(500).json({ message: 'Unable to update your post.' });
+    next(new AppError(500, 'Unable to update your post', error));
   }
 };
 
@@ -69,11 +71,7 @@ const deletePost: RequestHandler = async (req, res, next) => {
   try {
     const { postId: _id } = req.params;
     const post = await Post.findOne({ _id, creator: req.user });
-
-    if (!post) {
-      res.status(404).json({ message: 'Post not found.' });
-      return;
-    }
+    if (!post) return next(new AppError(404, 'Post not found'));
 
     if (post.imgURL) deleteFile(post.imgURL);
     await Post.deleteOne({ _id, creator: req.user });
@@ -81,8 +79,7 @@ const deletePost: RequestHandler = async (req, res, next) => {
     io.emit(`post:${_id}:delete`, post); // emits to specfic path only
     res.status(200).json(null); // 200 replaces client Data, so post must be null
   } catch (error) {
-    captainsLog(5, 'deletePost Catch', error);
-    res.status(500).json({ message: 'Unable to delete post.' });
+    next(new AppError(500, 'Unable to delete your post', error));
   }
 };
